Module classes splitted
